Interface and Impl service added / method 'sendEmail' refactored
