+++ src/main/java/com/alkemy/ong/service/IEmailService.java
@@ -1,0 +1,9 @@
+package com.alkemy.ong.service;
+
+import com.sendgrid.Response;
+
+public interface IEmailService {
+    
+    Response sendEmail(String toEmail);
+
+}
